build: switch from tsup to tsdown
Replace tsup with tsdown (rolldown-based) for building the js-sdk and cli
packages. Published artifact layout is preserved: js-sdk ships
dist/index.js (CJS), dist/index.mjs (ESM) and .d.ts/.d.mts; the cli ships
an executable dist/index.js plus dist/templates.

The cli build targets Node 20, and engines.node for both packages is set
to ">=20.18.1 <21 || >=22" (keeps the minimum required by undici while
excluding the end-of-life Node 21 line). Pin the nodejs in .tool-versions
to 22.18.0 (satisfies tsdown's and oxlint's engine requirements).
