+++ .tool-versions
@@ -1,5 +1,5 @@
 deno 1.46.3
-nodejs 20.19.5
+nodejs 22.18.0
 pnpm 9.15.5
 python 3.10
 uv 0.10.0
